+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - [V2] Add support for MaglevHashEndpoints
 - [V2] Add basic support for Views and Analyzers
 - [V2] Add ServerMode/SetServerMode/ServerID
-- [V2] Add collection Truncate, Properties, SetProperties
+- [V2] Add collection Truncate, Count, Properties, SetProperties
 - [V2] Add and re-organize missing collection properties fields
 
 ## [1.6.0](https://github.com/arangodb/go-driver/tree/v1.6.0) (2023-05-30)